moves login api to login component
