+++ src/main/frontend/src/app/app.middleware.js
@@ -1,6 +1,6 @@
 import { call, put, takeEvery } from 'redux-saga/effects';
 import { actionCreator, actionFailed } from 'shared/actionHelper';
-import { fetchMicroservicesApi, loginApi, fetchAvailableStagesApi } from './app.api';
+import { fetchMicroservicesApi, fetchAvailableStagesApi } from './app.api';
 
 
 export function* fetchAvailableStages(action) {
@@ -9,16 +9,6 @@ export function* fetchAvailableStages(action) {
     yield put(actionCreator('FETCH_AVAILABLE_STAGES_SUCCESS', user));
   } catch (e) {
     yield put(actionFailed('FETCH_AVAILABLE_STAGES_FAILED', e));
-    yield put(actionCreator('OPEN_SNACKBAR', 'Stages konnten nicht geladen werden'));
-  }
-}
-
-export function* login(action) {
-  try {
-    const user = yield call(loginApi, action.payload);
-    yield put(actionCreator('LOGIN_SUCCESS', user));
-  } catch (e) {
-    yield put(actionFailed('LOGIN_FAILED', e));
     yield put(actionCreator('OPEN_SNACKBAR', 'Stages konnten nicht geladen werden'));
   }
 }
@@ -36,7 +26,6 @@ export function* fetchMicroservices(action) {
 function* appSaga() {
   yield takeEvery('FETCH_AVAILABLE_STAGES_REQUESTED', fetchAvailableStages);
   yield takeEvery('FETCH_MICROSERVICES_REQUESTED', fetchMicroservices);
-  yield takeEvery('LOGIN_REQUESTED', login);
 }
 
 export default appSaga;
